+++ docs/release-notes.rst
@@ -36,6 +36,7 @@ Fixed PhantomJS conformance issues
 - After creating a ``webpage``, evaluating javascript should be possible.
   The window is now opened immediately with about:blank when calling create().
 - ``webpage.onResourceRequested`` should receive a request object as second parameter
+- ``webpage.setContent()`` should be able to receive a DOM element, not only a string 
 
 version 0.7
 ===========